# docs/openapi/openapi-drapt.yaml
openapi: 3.1.0    # Sửa lại 
info:
  title: Health Tracking API
  version: 0.1.0
  description: >
    Draft OpenAPI contract aligned with the project's flowcharts and database models.
    Authentication: Bearer access token. Most write/read endpoints require auth.
servers:
  - url: /api/v1
paths:
...
  /exercises/search:
    get:
      tags: [Exercises]
      summary: Search exercises (returns global + user's custom)
      operationId: exercises_search
      security:
        - HTTPBearer: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: major_heading
          in: query
          required: false
          schema:
            type: string
            nullable: true
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            nullable: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseSearchResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /exercises:
    post:
      tags: [Exercises]
      summary: Create a custom exercise for the current user
      operationId: exercises_create
      security:
        - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /exercises/{exercise_id}:
    get:
      tags: [Exercises]
      summary: Get exercise detail
      operationId: exercises_get
      security:
        - HTTPBearer: []
      parameters:
        - name: exercise_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Exercise not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Exercises]
      summary: Update a custom exercise (owner only)
      operationId: exercises_patch
      security:
        - HTTPBearer: []
      parameters:
        - name: exercise_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExercisePatchRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not allowed (not owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Exercise not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags: [Exercises]
      summary: Delete a custom exercise (soft delete; owner only)
      operationId: exercises_delete
      security:
        - HTTPBearer: []
      parameters:
        - name: exercise_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not allowed (not owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Exercise not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /logs/exercises:
    get:
      tags: [Logs]
      summary: List exercise logs for a day
      operationId: logs_exercises_list
      security:
        - HTTPBearer: []
      parameters:
        - name: day
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseDiaryDayResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items: {}
        msg:
          type: string
        type:
          type: string
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'

    # ---- Shared ----
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
...
    ExerciseLogItem:
      type: object
      required: [id, entry_id, exercise_id, duration_min]
      properties:
        id:
          type: integer
          format: int64
        entry_id:
          type: integer
          format: int64
        exercise_id:
          type: integer
          format: int64
        duration_min:
          type: number
        met_value_snapshot:
          type: number
          nullable: true
        calories:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
    ExerciseLogEntry:
      type: object
      required: [id, logged_at]
      properties:
        id:
          type: integer
          format: int64
        logged_at:
          type: string
          format: date-time
        total_calories:
          type: number
          nullable: true
    ExerciseDiaryDayResponse:
      type: object
      required: [day, entry, items]
      properties:
        day:
          type: string
          format: date
        entry:
          $ref: '#/components/schemas/ExerciseLogEntry'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseLogItem'

    # ---- Biometrics ----
...